# hotel-booking-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Booking API",
        "description": "REST API cho hệ thống đặt phòng khách sạn.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/dat-phong-cua-toi": {
            "get": {
                "tags": [
                    "Đặt phòng"
                ],
                "summary": "Lấy danh sách đặt phòng của tôi",
                "description": "Chỉ trả về các đặt phòng thuộc người dùng đang được xác thực bởi Bearer token Sanctum.",
                "operationId": "9309a3ad924047d98a071129bf6137d8",
                "parameters": [
                    {
                        "name": "phan_loai_luu_tru",
                        "in": "query",
                        "description": "Lọc theo trạng thái lưu trú",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "sap_toi",
                                "dang_luu_tru",
                                "da_hoan_thanh",
                                "da_huy"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy danh sách đặt phòng thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy danh sách đặt phòng của bạn thành công"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "ma_dat_phong": {
                                                        "type": "string",
                                                        "example": "DP202608190001"
                                                    },
                                                    "ten_phong": {
                                                        "type": "string",
                                                        "example": "Deluxe 201"
                                                    },
                                                    "hinh_anh": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "ngay_nhan_phong": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "ngay_tra_phong": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "so_luong_khach": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "tong_tien": {
                                                        "type": "number",
                                                        "example": 1600000
                                                    },
                                                    "tien_phong": {
                                                        "type": "number",
                                                        "example": 1400000
                                                    },
                                                    "tien_dich_vu": {
                                                        "type": "number",
                                                        "example": 200000
                                                    },
                                                    "dich_vu": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "ten_dich_vu": {
                                                                    "type": "string",
                                                                    "example": "Ăn sáng"
                                                                },
                                                                "so_luong": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "don_gia": {
                                                                    "type": "number",
                                                                    "example": 100000
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "trang_thai": {
                                                        "type": "string",
                                                        "example": "cho_xac_nhan"
                                                    },
                                                    "phuong_thuc_thanh_toan": {
                                                        "type": "string",
                                                        "example": "tai_khach_san"
                                                    },
                                                    "trang_thai_thanh_toan": {
                                                        "type": "string",
                                                        "example": "chua_thanh_toan"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Chưa xác thực"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/dat-phong-cua-toi/{id}": {
            "get": {
                "tags": [
                    "Đặt phòng"
                ],
                "summary": "Lấy chi tiết đặt phòng của tôi",
                "description": "Chỉ trả về đặt phòng có ID thuộc người dùng đang được xác thực bởi Bearer token Sanctum.",
                "operationId": "9e9cc78505e137adfe9b0ed157660759",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy chi tiết đặt phòng thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy chi tiết đặt phòng thành công"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "ma_dat_phong": {
                                                    "type": "string",
                                                    "example": "DP202608190001"
                                                },
                                                "ten_phong": {
                                                    "type": "string",
                                                    "example": "Deluxe 201"
                                                },
                                                "hinh_anh": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "ten_khach_hang": {
                                                    "type": "string",
                                                    "example": "Nguyễn Văn A"
                                                },
                                                "ngay_nhan_phong": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "ngay_tra_phong": {
                                                    "type": "string",
                                                    "format": "date"
                                                },
                                                "so_dem": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "tong_tien": {
                                                    "type": "number",
                                                    "example": 1600000
                                                },
                                                "ghi_chu": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "trang_thai": {
                                                    "type": "string",
                                                    "example": "cho_xac_nhan"
                                                },
                                                "phuong_thuc_thanh_toan": {
                                                    "type": "string",
                                                    "example": "tai_khach_san"
                                                },
                                                "trang_thai_thanh_toan": {
                                                    "type": "string",
                                                    "example": "chua_thanh_toan"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Chưa xác thực"
                    },
                    "404": {
                        "description": "Không tìm thấy đặt phòng"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/dat-phong-cua-toi/{id}/huy": {
            "put": {
                "tags": [
                    "Đặt phòng"
                ],
                "summary": "Hủy đặt phòng của tôi",
                "description": "Chỉ cho phép hủy đặt phòng thuộc người dùng đang xác thực và còn ở trạng thái chờ xác nhận.",
                "operationId": "c95ddcbc75931f1c3ff16732acf22fe4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hủy đặt phòng thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Hủy đặt phòng thành công."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "trang_thai": {
                                                    "type": "string",
                                                    "example": "da_huy"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Chưa xác thực"
                    },
                    "404": {
                        "description": "Không tìm thấy đặt phòng"
                    },
                    "422": {
                        "description": "Đặt phòng không thể hủy"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Bearer Token",
                "scheme": "bearer"
            }
        }
    },
...
}